refactor: remove CMS, hardcode portfolio data as static JSON
Drops Firebase/Firestore, Cloudinary uploads, and the /login + /profile
CMS dashboard. Content now lives in data/portfolio.json and is imported
directly via lib/portfolio.ts. Adds public/certificates/ for certificate
image files (referenced via a certificate's optional `image` field).
