--- lib/storage.ts
@@ -1,22 +1,0 @@
-async function uploadToCloudinary(file: File, publicId: string): Promise<string> {
-  const cloudName = process.env.NEXT_PUBLIC_CLOUDINARY_CLOUD_NAME;
-  const uploadPreset = process.env.NEXT_PUBLIC_CLOUDINARY_UPLOAD_PRESET;
-  if (!cloudName || !uploadPreset) {
-    throw new Error("Cloudinary not configured. Add NEXT_PUBLIC_CLOUDINARY_CLOUD_NAME and NEXT_PUBLIC_CLOUDINARY_UPLOAD_PRESET to env.");
-  }
-  const form = new FormData();
-  form.append("file", file);
-  form.append("upload_preset", uploadPreset);
-  form.append("public_id", publicId);
-  const res = await fetch(`https://api.cloudinary.com/v1_1/${cloudName}/image/upload`, { method: "POST", body: form });
-  if (!res.ok) {
-    const err = await res.json().catch(() => ({}));
-    throw new Error((err as { error?: { message?: string } }).error?.message ?? "Upload failed");
-  }
-  const data = await res.json() as { secure_url: string };
-  return data.secure_url;
-}
-
-export async function uploadProfilePhoto(file: File): Promise<string> {
-  return uploadToCloudinary(file, `profile_${Date.now()}`);
-}
