+++ package.json
@@ -9,7 +9,6 @@
     "lint": "eslint ."
   },
   "dependencies": {
-    "firebase": "^12.15.0",
     "next": "^16.1.1",
     "react": "^19.2.3",
     "react-dom": "^19.2.3"
